[JRO] Clean up dependencies and get building
There were some weird errors popping up when building the project so
these have been updated - both with the gems and the npm packages.

The port for browsersync has been updated to localhost:5000. This is a port not
taken by anyone as a domain on shubox.io - so we can hook into it.

Jekyll now builds with "bundle exec" to avoid the "Unresolved specs
during Gem::Specification.reset" error.