[JRO] Make codepen demos more robust w/base CSS
CodePen gives you a nice blank canvas, which is nice, but to add halfway
decent boilerplate CSS per each "pen" would be too much work. To that
end they provide a way to link off to external CSS (or JS) to use in
your demo in addition to the demo-specific code.

This commit

* creates a base "codepen" css file we can use specifically for those demos.
* creates a new gulp task called "codepen" where, if provided the
  requisite ENV vars, it will upload the css file to the correct S3
  bucket whenever the file is changed.

NOTE: The address to the CSS file is:

> http://shubox-codepen-io.s3.amazonaws.com/codepen.css