Add .editorconfig file
There's an interesting project here: http://editorconfig.org/ ... that
attempts to define some baseline editor settings that can be used across
many different code editors. See: http://editorconfig.org/#download

The purpose of integrating into shubox projects it to make sure we're
all working with the same assumptions and standards for things like:

* indentation (tabs v spaces)
* width of tab (x number of spaces)
* character encoding / charset
* newlines at end of file (or not)
* etc

I've noticed that some HTML files are indented with 4 character widths,
some with 2, some with tabs, some with spaces. It'd be beneficial to us
to agree with one and stick with it.

NOTE: this is a proposal in PR form. This is not a mandate.

Dan, what do you think?