Able to deploy to S3 bucket!
(Deployable, not to be confused with deplorable. Nothing similar to the
latter here.)

A website don't mean jack if it doesn't make it up to "THE CLOUD" (nee
"the internet") ... so let's get something together that allows us to
deploy it!

```sh
AWS_ACCESS_KEY=##### AWS_SECRET_KEY=##### S3_BUCKET=bucket-name gulp deploy
```

Running that with the necessary credentials will take all the files in
./public and will upload them to the s3 bucket where the files will
live.

Talk to @jayroh about what those ENV variables need to be in order for
those to get into staging and production buckets.