Add more, and better, documentation.
Image transforms was conspicuously absent from all documentation and in
desperate need of some representation. This commit adds some info around
how to set it up and what is provided.

Also added are some notes about the new sandbox feature, some fixes to
the content that is there currently, and fixing the script tags pointing
to the highlight and jquery scripts.